pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rainbot"
version = "2.5.3"
description = "A powerful Discord moderation bot with advanced automod and logging features"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "rainbot team"}
]
keywords = ["discord", "bot", "moderation", "automod", "logging"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "discord.py~=2.5.3",
    "motor~=3.3.2",
    "dnspython~=2.4.2",
    "python-dotenv~=1.0.0",
    "parsedatetime~=2.6",
    "python-dateutil~=2.8.2",
    "tensorflow~=2.15.0",
    "nudenet~=2.0.7",
    "imagehash~=4.3.1",
    "cachetools~=5.3.2",
    "emoji~=2.8.0",
    "aiofiles~=23.2.1",
    "colorama~=0.4.6",
    "rich~=13.7.0",
    "typing-extensions~=4.9.0",
]

[project.optional-dependencies]
dev = [
    "black==23.12.1",
    "flake8==7.0.0",
    "isort==5.13.2",
    "mypy==1.8.0",
    "pylint==3.0.3",
    "pytest==7.4.4",
    "pytest-asyncio==0.23.5",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "sphinx==7.2.6",
    "sphinx-rtd-theme==2.0.0",
    "myst-parser==2.0.0",
    "pre-commit==3.6.0",
    "tox==4.12.1",
    "coverage==7.4.1",
    "ipython==8.18.1",
    "ipdb==0.13.13",
    "types-requests==2.31.0.20231225",
    "types-PyYAML==6.0.12.12",
    "memory-profiler==0.61.0",
    "line-profiler==4.1.2",
    "bandit==1.7.5",
    "safety==2.3.5",
]

[project.urls]
Homepage = "https://github.com/fourjr/rainbot"
Documentation = "https://github.com/fourjr/rainbot/wiki"
Repository = "https://github.com/fourjr/rainbot"
"Bug Tracker" = "https://github.com/fourjr/rainbot/issues"
"Support Server" = "https://discord.gg/eXrDpGS"

[project.scripts]
rainbot = "bot:main"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = ["bot", "cogs", "ext"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "discord.*",
    "motor.*",
    "tensorflow.*",
    "nudenet.*",
    "imagehash.*",
    "parsedatetime.*",
    "emoji.*",
    "rich.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["bot", "cogs", "ext"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/stubs/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "stubs"]
skips = ["B101", "B601"]

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
    "R0914",  # too-many-locals
    "W0621",  # redefined-outer-name
    "W0622",  # redefined-builtin
]

[tool.pylint.format]
max-line-length = 100

[tool.pylint.design]
max-args = 10
max-locals = 20
max-returns = 10
max-branches = 15
max-statements = 60
max-parents = 7
max-attributes = 10
min-public-methods = 0
max-public-methods = 20

[tool.pylint.similarities]
min-similarity-lines = 4
ignore-comments = "yes"
ignore-docstrings = "yes"
ignore-imports = "yes" 